完成了文章的增删改查，准备接入前端框架

+++ src/main/java/com/yan/ssm/dao/ArticleDao.java
@@ -17,11 +17,11 @@ public interface ArticleDao {
 
     int updateSelective(Article article);
 
-    int deleteById(int id);
+    int deleteById(Long id);
 
     List<Article> selectByParams (Map<String,Object> params);
 
-    Article selectByPrimaryKey(int id);
+    Article selectByPrimaryKey(Long id);
 
     List<Article> selectAllArticle();
 }
